fixing #184
