--- src/components/project/month.const.js
@@ -1,31 +1,0 @@
-export const month = (m) => {
-    var matt = ''
-    switch(m){
-        case 1: matt = 'January'
-            break;
-        case 2: matt = 'February'
-            break;
-        case 3: matt = 'March'
-            break;
-        case 4: matt = 'April'
-            break;
-        case 5: matt = 'May'
-            break;
-        case 6: matt = 'June'
-            break;
-        case 7: matt = 'July'
-            break;
-        case 8: matt = 'August'
-            break;
-        case 9: matt = 'September'
-            break;
-        case 10: matt = 'October'
-            break;
-        case 11: matt = 'November'
-            break;
-        case 12: matt = 'December'
-            break;
-        default:
-    }
-    return matt
-}
